% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation_batch.R
\name{add_input_survey_cols}
\alias{add_input_survey_cols}
\title{Add input survey columns}
\usage{
add_input_survey_cols(data, survey_colnames)
}
\arguments{
\item{data}{the data to add input survey cols}

\item{survey_colnames}{the survey columns that will be stored}
}
\value{
dataframe/tibble with survey columns
}
\description{
add empty survey input columns 
and its corresponding filepath, and empty timestamps
to fill out in shiny session
}
