% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_helpers.R
\name{attempt_login}
\alias{attempt_login}
\title{Attempt to log into Synapse}
\usage{
attempt_login(syn, ...)
}
\arguments{
\item{syn}{Synapse client object}

\item{...}{Synapse credentials, such as `authToken` or `email` with a
`password` or `apiKey`.}
}
\description{
Attempt to log into Synapse. Will first try using authentication
credentials written to a .synapseConfig file. If that fails, will try using
any credentials passed to the function. Will return `NULL` if not all
attempts failed.
}
