% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_filehandles.R
\name{batch_process_filehandles}
\alias{batch_process_filehandles}
\title{batch_process_filehandle Server Function}
\usage{
batch_process_filehandles(
  syn,
  values,
  synapse_tbl,
  filehandle_cols,
  uid,
  survey_colnames,
  keep_metadata,
  n_batch
)
}
\description{
batch_process_filehandle Server Function
}
