% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_table_column_files.R
\name{batch_process_table_column_files}
\alias{batch_process_table_column_files}
\title{Process Synapse Table Column Files}
\usage{
batch_process_table_column_files(
  syn,
  all_data,
  curated_data,
  synapse_tbl_id,
  filehandle_cols,
  uid,
  survey_colnames,
  keep_metadata,
  n_batch,
  sort_keys,
  output_location,
  cache_location,
  visualization_funs
)
}
\arguments{
\item{syn}{synapseclient}

\item{all_data}{dataframe of whole data from synapse table}

\item{curated_data}{dataframe of annotated ata}

\item{synapse_tbl_id}{synapse source table id}

\item{filehandle_cols}{filehandle column target to parse}

\item{uid}{unique identifier of each files}

\item{keep_metadata}{metadata to keep from the table}

\item{n_batch}{number of batch per session}

\item{output_location}{where to store processed files}

\item{cache_location}{where to find raw files}

\item{visualization_funs}{function to visualize data}
}
\value{
a dataframe containing processed 
Table column files that will be used
for rendering in the Shiny App
}
\description{
Process Synapse Table Filehandles in batch processing,
will take unannotated data and batch process each 
of them based on a visualization function
}
