% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{check_survey_config}
\alias{check_survey_config}
\title{Get Survey Config}
\usage{
check_survey_config(config)
}
\arguments{
\item{config}{config}
}
\value{
config file/error messages if config file is invalid
}
\description{
This is a helper to validate shiny App survey buttons 
configurations and see whether user has the 
right inputs for the Shiny App
}
