% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_directory.R
\name{clear_user_directory}
\alias{clear_user_directory}
\title{Function to clear user directory
storing image files (to not overpopulate Shiny Server)}
\usage{
clear_user_directory(curr_annotator)
}
\arguments{
\item{curr_annotator}{user name of the annotator}
}
\description{
Function to clear user directory
storing image files (to not overpopulate Shiny Server)
}
