% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_user_directory}
\alias{create_user_directory}
\title{Function to create user directory for temporarily
storing image files (to not overpopulate Shiny Server)}
\usage{
create_user_directory(curr_annotator)
}
\arguments{
\item{curr_annotator}{user name of the annotator
create user directory}
}
\description{
Function to create user directory for temporarily
storing image files (to not overpopulate Shiny Server)
}
