% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation_batch.R
\name{get_new_rows}
\alias{get_new_rows}
\title{Get new rows (unannotated)}
\usage{
get_new_rows(data, curated_data, uid)
}
\arguments{
\item{data}{data}

\item{curated_data}{curated data stored in synapse}

\item{uid}{unique identifier to compare whether image
has been annotated or not}
}
\value{
dataframe/tibble with unannotated rows
}
\description{
get unannotated rows by anti-joining
fetched data with previously curated data
}
