% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_images.R
\name{get_session_images}
\alias{get_session_images}
\title{Get unannotated files in batch}
\usage{
get_session_images(
  data,
  syn,
  synapse_tbl_id,
  filehandle_cols,
  uid,
  n_batch,
  cache_location,
  keep_metadata = NULL
)
}
\arguments{
\item{data}{containing un-annotated data based on each user}

\item{syn}{synapse object}

\item{synapse_tbl_id}{synapse table id}

\item{filehandle_cols}{table filehandles}

\item{uid}{unique ID of images/plot/graphs}

\item{n_batch}{number of images/plot/graphs per annotation session}

\item{cache_location}{where to cache each user annotations}

\item{keep_metadata}{metadata to keep}
}
\value{
unannotated data based on previously stored records
}
\description{
Helper function for downloading a number of Synapse table-attached 
files according to input batch, and process using desired function
}
