% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_table.R
\name{get_source_table}
\alias{get_source_table}
\title{Get source table}
\usage{
get_source_table(syn, filehandle_cols, synapse_tbl_id)
}
\arguments{
\item{syn}{synapseclient}

\item{filehandle_cols}{column filehandles}

\item{synapse_tbl_id}{synapse ID of Table}
}
\value{
dataframe of tidy source synapse table
}
\description{
Get source table from synapse and pivot
table into tidy format based on filehandle
columns
}
