% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_images.R
\name{get_table_string_filters}
\alias{get_table_string_filters}
\title{Get Table Unique Identifier String Filter}
\usage{
get_table_string_filters(uid)
}
\arguments{
\item{uid}{unique identifier used to}
}
\value{
a string of unique identifier that will
be included in the batch with parentheses
}
\description{
This is a helper function to build string filter (SQL-like)
for filtering Synapse Table before downloading files 
Note: This is done to enable small-batch download
}
