% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_table_column_files.R
\name{get_unannotated_files_in_batch}
\alias{get_unannotated_files_in_batch}
\title{Get unannotated files in batch}
\usage{
get_unannotated_files_in_batch(
  syn,
  data,
  synapse_tbl_id,
  filehandle_cols,
  uid,
  keep_metadata,
  n_batch,
  output_location,
  cache_location,
  visualization_funs
)
}
\arguments{
\item{data}{containing un-annotated data based on each user}
}
\description{
Helper function for downloading a number of Synapse table-attached 
files according to input batch, and process using desired function
}
