% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{mod_main_server}
\alias{mod_main_server}
\title{Server module for main function}
\usage{
mod_main_server(id, syn)
}
\arguments{
\item{id}{namespace identifier}

\item{syn}{user Synapseclient object
get all data and previous data
get previous image that has been curated}
}
\description{
create backend process of the RShiny App
}
