% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{mod_main_ui}
\alias{mod_main_ui}
\title{UI Function main function}
\usage{
mod_main_ui(id)
}
\arguments{
\item{id}{identifier namespace}
}
\value{
UI a dashboard page object that contains the RShiny App UI
}
\description{
creates the RShiny overall UI
}
