% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_input_user.R
\name{mod_survey_input_user_server}
\alias{mod_survey_input_user_server}
\title{select_view Server Function}
\usage{
mod_survey_input_user_server(input, output, session, values)
}
\description{
select_view Server Function
}
