% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_image_opts}
\alias{parse_image_opts}
\title{Function to parse image options}
\usage{
parse_image_opts(config)
}
\arguments{
\item{config}{config}
}
\value{
named list of image options being used
}
\description{
Function to parse image options
}
