% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_initial_selection}
\alias{parse_initial_selection}
\title{function to parse initial selection
place null if user is not giving any input}
\usage{
parse_initial_selection(input)
}
\arguments{
\item{input}{user input}
}
\description{
function to parse initial selection
place null if user is not giving any input
}
