% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inputs.R
\name{parse_slider_value}
\alias{parse_slider_value}
\title{Function to parse slider (single/multiple)}
\usage{
parse_slider_value(row_input, selected)
}
\arguments{
\item{row_input}{input from user}

\item{selected}{previously selected}
}
\value{
the selected shiny user input
}
\description{
Function to parse slider (single/multiple)
}
