% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_survey_opts.R
\name{parse_survey_opts}
\alias{parse_survey_opts}
\title{function to parse survey question from the config file}
\usage{
parse_survey_opts(config)
}
\arguments{
\item{config}{annotator config file being used}
}
\value{
named list of survey question and its options
instantiate survey configuration
read survey configuratation
read survey configuration
}
\description{
function to parse survey question from the config file
}
