% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_synapse_opts}
\alias{parse_synapse_opts}
\title{function to parse synapse options}
\usage{
parse_synapse_opts(config)
}
\arguments{
\item{config}{config}
}
\value{
named list of synapse options being used
instantiate empty list
read synapse configuratiton 
return list of values
}
\description{
function to parse synapse options
}
