% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_uid_to_string}
\alias{parse_uid_to_string}
\title{Function to parse unique id for iteration
THIS IS MODULAR}
\usage{
parse_uid_to_string(uid)
}
\description{
Function to parse unique id for iteration
THIS IS MODULAR
}
