% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_render_image.R
\name{mod_render_image_ui}
\alias{mod_render_image_ui}
\alias{mod_render_image_server}
\title{UI Function for the image rendering}
\usage{
mod_render_image_ui(id)

mod_render_image_server(
  input,
  output,
  session,
  obj_path,
  input_width = 650,
  input_height = 450
)
}
\arguments{
\item{id}{the id}

\item{input}{the input variables from [shiny::callModule()]}

\item{output}{the output variables from [shiny::callModule()]}

\item{session}{the session from [shiny::callModule()]}

\item{obj_path}{image filepath}

\item{input_width}{image width}

\item{input_height}{image height}
}
\value{
UI for the image rendering for the dashboard


}
\description{
Creates the UI for image rendering

Creates the server for image rendering, complete with 
customizable image width/height
}
