% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation_batch.R
\name{sort_rows}
\alias{sort_rows}
\title{Sort rows arbitrary to sorting keys}
\usage{
sort_rows(data, sort_keys)
}
\arguments{
\item{data}{the data to sort on}

\item{sort_keys}{the keys to sort based off}
}
\value{
sorted dataframe/tibble
}
\description{
sort dataframe rows for annotations
based on the keys provided
}
