% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_inputs.R
\name{store_inputs}
\alias{store_inputs}
\title{Function to store survey inputs from shiny app}
\usage{
store_inputs(data, curr_index, user_inputs, keep_metadata, uid)
}
\arguments{
\item{data}{data of annotations}

\item{curr_index}{current index marker}

\item{user_inputs}{input given by users}

\item{keep_metadata}{what metadata to keep}

\item{uid}{what are the unique ids (recordId, healthcode, participantId etc.)}
}
\value{
dataframe of annotation based on current index
}
\description{
Function to store survey inputs from shiny app
}
