% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tbl_to_synapse.R
\name{store_tbl_to_synapse}
\alias{store_tbl_to_synapse}
\title{Function to store annotator's table to synapse}
\usage{
store_tbl_to_synapse(
  syn,
  synapseclient,
  parent_id,
  new_data,
  stored_data,
  current_annotator,
  output_filename,
  ...
)
}
\arguments{
\item{syn}{synapse client object}

\item{synapseclient}{synapseclient library}

\item{parent_id}{parent id to store dataframe}

\item{new_data}{current annotations dataframe}

\item{stored_data}{previously stored annotationx dataframe}

\item{current_annotator}{current annotator as file prefix}

\item{output_filename}{name of the data output}

\item{...}{additional info, will be used for provenance}
}
\value{

}
\description{
Function to store annotator's table to synapse
}
