% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_input_store.R
\name{survey_input_store}
\alias{survey_input_store}
\title{Function to store survey inputs from shiny app}
\usage{
survey_input_store(
  data,
  curr_index,
  user_inputs,
  survey_colnames,
  keep_metadata,
  uid
)
}
\arguments{
\item{data}{data of annotations}

\item{curr_index}{current index marker}

\item{user_inputs}{input given by users}

\item{survey_colnames}{columns used for the surveys}

\item{keep_metadata}{what metadata to keep}

\item{uid}{what are the unique ids (recordId, healthcode, participantId etc.)}
}
\value{
dataframe of annotation based on current index
}
\description{
Function to store survey inputs from shiny app
}
