% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_input_user.R
\name{mod_survey_input_user_server}
\alias{mod_survey_input_user_server}
\title{Server Function for creating survey buttons}
\usage{
mod_survey_input_user_server(input, output, session, survey_colnames, values)
}
\arguments{
\item{input}{the input variables from [shiny::callModule()]}

\item{output}{the output variables from [shiny::callModule()]}

\item{session}{the session from [shiny::callModule()]}

\item{survey_colnames}{the column used to store survey}

\item{values}{the reactive values from server}
}
\value{
the UI for survey inputs in ShinyApp
change input
}
\description{
Gives functionality to creating the survey buttons server-side,
helps parse through each survey columns requirement and render buttons
according to user configuration. 
Designed to take multiple input as 'comma-separated'
}
