% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_input_user.R
\name{mod_survey_input_user_ui}
\alias{mod_survey_input_user_ui}
\title{UI Function for creating survey buttons}
\usage{
mod_survey_input_user_ui(id)
}
\arguments{
\item{id}{the id}
}
\value{
the UI for survey inputs in ShinyApp
read config file
create based on config
set button
}
\description{
Gives functionality to creating the survey buttons UI,
populating with menu that can run different kinds of survey buttons
based on different input types
}
