% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_input_update.R
\name{update_buttons}
\alias{update_buttons}
\title{Function to update user input shiny buttons by
maintaining previously selected buttons and/or
updating it with new ones}
\usage{
update_buttons(reactive_values, session, curr_index, config)
}
\arguments{
\item{reactive_values}{access reactive values}

\item{session}{access shiny session}

\item{curr_index}{access current index}

\item{config}{access annotator config file}
}
\value{

}
\description{
Function to update user input shiny buttons by
maintaining previously selected buttons and/or
updating it with new ones
}
