% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inputs.R
\name{update_inputs}
\alias{update_inputs}
\title{Function to update user input shiny buttons by
maintaining previously selected buttons and/or
updating it with new ones}
\usage{
update_inputs(reactive_values, session, curr_index, config)
}
\arguments{
\item{reactive_values}{access reactive values}

\item{session}{access shiny session}

\item{curr_index}{access current index}

\item{config}{access annotator config file}
}
\value{
dataframe with updated inputs
}
\description{
Function to update user input shiny buttons by
maintaining previously selected buttons and/or
updating it with new ones
}
