% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config_file.R
\name{validate_config_file}
\alias{validate_config_file}
\title{Validate Config File}
\usage{
validate_config_file(config_path)
}
\arguments{
\item{config_path}{path to configuration file}
}
\value{
TRUE if config file is already in the right format
}
\description{
This is a function to validate the config file 
that you will parse for the Shiny App
}
