% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_table_column_files.R
\name{visualize_column_files}
\alias{visualize_column_files}
\title{Visualize Synapse Table Column Files}
\usage{
visualize_column_files(data, funs, output_location)
}
\arguments{
\item{data}{data where it contains cached `filePath` of the table attached files}

\item{funs}{custom visualization function}

\item{output_location}{where to output the processed files location}
}
\value{
a dataframe containing processed files
}
\description{
Helper function to visualize synapse column files
based on a custom function
}
