% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{visualize_dvmotion_mpower_v1_sensors}
\alias{visualize_dvmotion_mpower_v1_sensors}
\title{function for mpower-V1 sensors visualization}
\usage{
visualize_dvmotion_mpower_v1_sensors(filepath)
}
\arguments{
\item{filepath}{the corresponding filepath being used}
}
\value{
the output filepath for the visualization
}
\description{
this function is used to visualize a 
filepath for gait data in Synapse table using mPower
V1 format (device_motion IOS) it will plot
acceleration, gyroscope and gravity readings on
x, y, z axis
}
