% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{visualize_mpower_gait}
\alias{visualize_mpower_gait}
\title{mpower gait visualization entry-point}
\usage{
visualize_mpower_gait(filepath)
}
\arguments{
\item{filepath}{the corresponding filepath being used}
}
\value{
the output filepath for the visualization
}
\description{
this function is used to visualize a 
filepath for gait data in Synapse table
}
