% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{visualize_mpower_v2_sensor}
\alias{visualize_mpower_v2_sensor}
\title{function for mpower-V2 sensors visualization}
\usage{
visualize_mpower_v2_sensor(filepath)
}
\arguments{
\item{filepath}{the corresponding filepath being used}
}
\value{
the output filepath for the visualization
}
\description{
this function is used to visualize a 
filepath of sensor in Synapse table using mPower
V2 format; it will plot acceleration, 
gyroscope and gravity readings on x, y, z axis
}
