% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{visualize_photo}
\alias{visualize_photo}
\title{function to parse in .png or .jpeg image}
\usage{
visualize_photo(filepath)
}
\arguments{
\item{filepath}{the corresponding filepath being used}
}
\value{
the output filepath for the visualization
}
\description{
this function is used to visualize a 
filepath for image Files in a Synapse table
}
