% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_images.R
\name{get_session_images}
\alias{get_session_images}
\title{Get unannotated files in batch}
\usage{
get_session_images(
  data,
  syn,
  synapse_tbl_id,
  filehandle_cols,
  uid,
  keep_metadata,
  n_batch,
  cache_location
)
}
\arguments{
\item{data}{containing un-annotated data based on each user}
}
\value{
unannotated data based on previously stored records
}
\description{
Helper function for downloading a number of Synapse table-attached 
files according to input batch, and process using desired function
}
