% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_table_column_files.R
\name{get_unannotated_files_in_batch}
\alias{get_unannotated_files_in_batch}
\title{Get unannotated files in batch}
\usage{
get_unannotated_files_in_batch(
  data,
  syn,
  synapse_tbl_id,
  filehandle_cols,
  uid,
  keep_metadata,
  n_batch,
  cache_location
)
}
\arguments{
\item{data}{containing un-annotated data based on each user}

\item{syn}{synapseclient}

\item{synapse_tbl_id}{synapse source table id}

\item{filehandle_cols}{filehandle column target to parse}

\item{uid}{unique identifier of each files}

\item{keep_metadata}{metadata to keep from the table}

\item{n_batch}{number of batch per session}

\item{cache_location}{where to find raw files}
}
\value{
unannotated data based on previously stored records
}
\description{
Helper function for downloading a number of Synapse table-attached 
files according to input batch, and process using desired function
}
