% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(config, funs)
}
\arguments{
\item{config}{config file to run your shiny}

\item{funs}{function to generate your images for the shiny app}
}
\description{
Run the Shiny Application
}
